Update mkl and mumps

+++ zscripts/uninstall-intel-mkl-and-ifort-debian.bash
@@ -2,14 +2,14 @@
 
 set -e
 
-VERSION="2023.2.0"
-
 # install packages
 sudo apt-get remove \
-    intel-oneapi-compiler-fortran-$VERSION \
-    intel-oneapi-mkl-$VERSION \
-    intel-oneapi-mkl-devel-$VERSION
+    intel-oneapi-compiler-fortran \
+    intel-oneapi-mkl \
+    intel-oneapi-mkl-devel
 
 # update ldconfig
-sudo rm /etc/ld.so.conf.d/intel-oneapi-mkl-and-compiler.conf
-sudo ldconfig
+if [ -f "/etc/ld.so.conf.d/intel-oneapi-mkl-and-compiler.conf" ]; then
+    sudo rm /etc/ld.so.conf.d/intel-oneapi-mkl-and-compiler.conf
+    sudo ldconfig
+fi
